Fix mangled metainfo file (#194)
* Fix mangled metainfo file

It turns out that add-provide subcommand of appstream-util mangles metainfo file which results in lost of important info like release changelogs.

The provide section doesn't look like something required so hopefully it can be dropped.

* Copy metadata file instead of renaming it

This will keep the original metadata file which will make easier to compare it with the modified one in order to spot regressions. 

The file will be ignored by flatpak as the name doesn't match appid.